+++ nc_test/tst_nofill.c
@@ -336,10 +336,10 @@ create_file(char *file_name, int fill_mode, size_t* sizehintp)
        {
 	   int izw;
 	   for(izw = 0; izw < TIME_LEN * LAT_LEN * LON_LEN; izw++) {
-	       zonal_wnd[izw] = 100 + i;
+               zonal_wnd[izw] = (float)(100 + i);
 	   }
 	   zonal_wnd_start[0] = 0;
-	   zonal_wnd_start[1] = i;
+	   zonal_wnd_start[1] = (size_t)i;
 	   zonal_wnd_start[2] = 0;
 	   zonal_wnd_start[3] = 0;
 	   zonal_wnd_count[0] = time_len;
@@ -370,7 +370,7 @@ main(int argc, char **argv)
     if (argc > 1) {
 	char *endptr, *str = argv[1];
 	errno = 0;
-	sizehint = strtol(str, &endptr, 0);
+	sizehint = (size_t)strtol(str, &endptr, 0);
 	/* check for various possible errors */
 	if ((errno == ERANGE && (sizehint == LONG_MAX || sizehint == LONG_MIN))
                    || (errno != 0 && sizehint == 0)) {
